+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/add_service/add_service_instruction.go
@@ -17,8 +17,6 @@ import (
 	"github.com/kurtosis-tech/stacktrace"
 	"go.starlark.net/starlark"
 	"go.starlark.net/starlarkstruct"
-	"regexp"
-	"strings"
 )
 
 const (
@@ -26,22 +24,7 @@ const (
 
 	serviceIdArgName     = "service_id"
 	serviceConfigArgName = "service_config"
-
-	serviceIdSubgroupName = "service_id"
-	allSubgroupName       = "all"
-	kurtosisNamespace     = "kurtosis"
-	// The placeholder format & regex should align
-	ipAddressReplacementRegex             = "(?P<" + allSubgroupName + ">\\{\\{" + kurtosisNamespace + ":(?P<" + serviceIdArgName + ">" + service.ServiceIdRegexp + ")\\.ip_address\\}\\})"
-	ipAddressReplacementPlaceholderFormat = "{{" + kurtosisNamespace + ":%v.ip_address}}"
-
-	unlimitedMatches = -1
-	singleMatch      = 1
-	subExpNotFound   = -1
 )
-
-// The compiled regular expression to do IP address replacements
-// Treat this as a constant
-var compiledRegex = regexp.MustCompile(ipAddressReplacementRegex)
 
 func GenerateAddServiceBuiltin(instructionsQueue *[]kurtosis_instruction.KurtosisInstruction, serviceNetwork service_network.ServiceNetwork) func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 	// TODO: Force returning an InterpretationError rather than a normal error
@@ -136,7 +119,7 @@ func (instruction *AddServiceInstruction) replaceIPAddress() error {
 	serviceIdStr := string(instruction.serviceId)
 	entryPointArgs := instruction.serviceConfig.EntrypointArgs
 	for index, entryPointArg := range entryPointArgs {
-		entryPointArgWithIPAddressReplaced, err := replaceIPAddressInString(entryPointArg, instruction.serviceNetwork, serviceIdStr)
+		entryPointArgWithIPAddressReplaced, err := shared_helpers.ReplaceIPAddressInString(entryPointArg, instruction.serviceNetwork, serviceIdStr)
 		if err != nil {
 			return stacktrace.Propagate(err, "Error occurred while replacing IP address in entry point args for '%v'", entryPointArg)
 		}
@@ -145,7 +128,7 @@ func (instruction *AddServiceInstruction) replaceIPAddress() error {
 
 	cmdArgs := instruction.serviceConfig.CmdArgs
 	for index, cmdArg := range cmdArgs {
-		cmdArgWithIPAddressReplaced, err := replaceIPAddressInString(cmdArg, instruction.serviceNetwork, serviceIdStr)
+		cmdArgWithIPAddressReplaced, err := shared_helpers.ReplaceIPAddressInString(cmdArg, instruction.serviceNetwork, serviceIdStr)
 		if err != nil {
 			return stacktrace.Propagate(err, "Error occurred while replacing IP address in command args for '%v'", cmdArg)
 		}
@@ -154,7 +137,7 @@ func (instruction *AddServiceInstruction) replaceIPAddress() error {
 
 	envVars := instruction.serviceConfig.EnvVars
 	for envVarName, envVarValue := range envVars {
-		envVarValueWithIPAddressReplaced, err := replaceIPAddressInString(envVarValue, instruction.serviceNetwork, serviceIdStr)
+		envVarValueWithIPAddressReplaced, err := shared_helpers.ReplaceIPAddressInString(envVarValue, instruction.serviceNetwork, serviceIdStr)
 		if err != nil {
 			return stacktrace.Propagate(err, "Error occurred while replacing IP address in env vars for '%v'", envVarValue)
 		}
@@ -162,30 +145,6 @@ func (instruction *AddServiceInstruction) replaceIPAddress() error {
 	}
 
 	return nil
-}
-
-func replaceIPAddressInString(originalString string, network service_network.ServiceNetwork, serviceIdForLogging string) (string, error) {
-	matches := compiledRegex.FindAllStringSubmatch(originalString, unlimitedMatches)
-	replacedString := originalString
-	for _, match := range matches {
-		serviceIdMatchIndex := compiledRegex.SubexpIndex(serviceIdSubgroupName)
-		if serviceIdMatchIndex == subExpNotFound {
-			return "", stacktrace.NewError("There was an error in finding the sub group '%v' in regexp '%v'. This is a Kurtosis Bug", serviceIdSubgroupName, compiledRegex.String())
-		}
-		serviceId := service.ServiceID(match[serviceIdMatchIndex])
-		ipAddress, found := network.GetIPAddressForService(serviceId)
-		if !found {
-			return "", stacktrace.NewError("'%v' depends on the IP address of '%v' but we don't have any registrations for it", serviceIdForLogging, serviceId)
-		}
-		ipAddressStr := ipAddress.String()
-		allMatchIndex := compiledRegex.SubexpIndex(allSubgroupName)
-		if allMatchIndex == subExpNotFound {
-			return "", stacktrace.NewError("There was an error in finding the sub group '%v' in regexp '%v'. This is a Kurtosis Bug", serviceIdSubgroupName, compiledRegex.String())
-		}
-		allMatch := match[allMatchIndex]
-		replacedString = strings.Replace(replacedString, allMatch, ipAddressStr, singleMatch)
-	}
-	return replacedString, nil
 }
 
 func (instruction *AddServiceInstruction) makeAddServiceInterpretationReturnValue() (*kurtosis_types.Service, *startosis_errors.InterpretationError) {
@@ -199,7 +158,7 @@ func (instruction *AddServiceInstruction) makeAddServiceInterpretationReturnValu
 			return nil, startosis_errors.NewInterpretationError("An error occurred while creating a port spec for values (number: '%v', port: '%v') the add instruction return value", portNumber, portProtocol)
 		}
 	}
-	ipAddress := starlark.String(fmt.Sprintf(ipAddressReplacementPlaceholderFormat, instruction.serviceId))
+	ipAddress := starlark.String(fmt.Sprintf(shared_helpers.IpAddressReplacementPlaceholderFormat, instruction.serviceId))
 	returnValue := kurtosis_types.NewService(ipAddress, portSpecsDict)
 	return returnValue, nil
 }