+++ docs/changelog.md
@@ -4,9 +4,11 @@
 - Add `wait` and `define` command in Startosis
 - Added `not found service GUIDs information` in `KurtosisContext.GetServiceLogs` method
 - Added a warning message in `service logs` CLI command when the request service GUID is not found in the logs database
+- Added ip address replacement in the JSON for `render_template` instruction
 
 ### Fixes
 - Fixes a bug where we'd propagate a nil error
+- Adds validation for `service_id` in `store_files_from_service`
 
 # 0.52.0
 ### Breaking Changes